+++ src/main/java/pro/zackpollard/telegrambot/api/internal/chat/message/MessageImpl.java
@@ -57,7 +57,7 @@ public class MessageImpl implements Message {
      * @return The ID of the message
      */
     @Override
-    public int getMessageId() {
+    public long getMessageId() {
         return message_id;
     }
